Little bit of cleanup, also started using C++11 nullptr instead of NULL which are wrong in places that expect a pointer to 0. Non FBO DepthCopy refined a bit, and fixed an attempt at turning of scissoring if model dimensions didnt fit.
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@7618 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ renderer/Profiling.cpp
@@ -134,7 +134,7 @@ private:
 
 	void AccumulateTotalTimes( section &s) {
 		s.totalCpuTimeMillis = s.totalGpuTimeMillis = 0;
-		s.count = s.queries.size();
+		s.count = (int)s.queries.size();
 		for ( auto& q : s.queries ) {
 			s.totalCpuTimeMillis += ( q.cpuStopTime - q.cpuStartTime ) * 1000 / Sys_ClockTicksPerSecond();
 			uint64_t gpuStartNanos, gpuStopNanos;
